rename: DailyExpenseRecord to ExpenseCard
