--- src/features/expense/ui/DailyExpenseRecord/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './DailyExpenseRecord';
