+++ src/features/expense/ui/ExpenseCard/ExpenseCard.test.tsx
@@ -1,16 +1,18 @@
 import { render } from '@testing-library/react';
 import { describe, it } from 'vitest';
 
+import { ConsumptionKind } from '@/features/expense/model/ConsumptionKind';
 import { Expense } from '@/features/expense/model/Expense';
 
-import DailyExpenseRecord from './DailyExpenseRecord';
+import ExpenseCard from './ExpenseCard';
 
-describe('DailyExpenseRecord', () => {
+describe('ExpenseCard', () => {
   const expenseProps: Omit<Expense, 'date'> = {
     uid: 'test-id',
     categories: [],
     amount: 1234567890,
     memo: 'description',
+    consumptionKind: ConsumptionKind.none,
   };
 
   const renderDailyExpenseRecord = ({
@@ -20,7 +22,7 @@ describe('DailyExpenseRecord', () => {
     memo,
   }: Omit<Expense, 'date'>) => {
     return render(
-      <DailyExpenseRecord
+      <ExpenseCard
         uid={uid}
         categories={categories}
         amount={amount}
