+++ src/features/expense/ui/ExpenseCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ExpenseCard';
